# openapi.yaml
openapi: 3.0.3
info:
  title: specCraft.ai Backend API
  description: |
    Express.js Backend API for specCraft.ai.
    This API generates software requirements specifications (SRS) and UML diagrams using Google Gemini and exports them to PDF/Markdown.
  version: 1.0.0
servers:
  - url: http://localhost:8001
    description: Local Development Server
paths:
...
  /api/projects:
    post:
      summary: Create a New Project
      description: Creates a new project workspace based on the user's wizard inputs. Restricted to a maximum of 3 project creations per user per day.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: Project created successfully from wizard input.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Project created successfully from wizard input
                  project:
                    $ref: '#/components/schemas/Project'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized. Missing or invalid Bearer JWT token.
        '403':
          description: Forbidden. Authenticated user does not have permission for this request.
        '429':
          description: Too Many Requests. Project creation limit (3 projects per day) or rate limit exceeded.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Project creation limit reached. You can only create up to 3 projects per day.
        '500':
          description: Internal server error.
    get:
      summary: List User Projects
      description: Returns a list of all projects owned by the currently authenticated user.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of projects retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  projects:
                    type: array
                    items:
                      $ref: '#/components/schemas/Project'
        '401':
          description: Unauthorized. Missing or invalid Bearer JWT token.
        '403':
          description: Forbidden. Access denied to user data.
        '429':
          description: Too Many Requests. Global rate limit exceeded.
  /api/projects/{id}:
    get:
      summary: Get Project Details
      description: Returns detailed metadata for a project, including its associated requirements specification.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Numeric project ID
          schema:
            type: integer
      responses:
        '200':
          description: Project details retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  project:
                    $ref: '#/components/schemas/Project'
                  specification:
                    $ref: '#/components/schemas/Specification'
        '401':
          description: Unauthorized. Missing or invalid Bearer JWT token.
        '403':
          description: Forbidden. User does not own the requested project.
        '404':
          description: Project not found.
        '429':
          description: Too Many Requests. Global rate limit exceeded.
    delete:
      summary: Delete a Project
      description: Deletes a project and all associated specification requirements (Cascading delete).
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Numeric project ID
          schema:
            type: integer
      responses:
        '200':
          description: Project deleted successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Project deleted successfully
        '401':
          description: Unauthorized. Missing or invalid Bearer JWT token.
        '403':
          description: Forbidden. User does not own the project.
        '404':
          description: Project not found.
        '429':
          description: Too Many Requests. Global rate limit exceeded.

  # Specifications Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Connect via Supabase Bearer Token JWT.
  schemas:
    Project:
      type: object
      properties:
        id:
          type: integer
          example: 12
        name:
          type: string
          example: Healthcare App
        description:
          type: string
          example: Peer-to-peer patient and doctor appointment booking platform.
        domain:
          type: string
          example: Healthcare
        complexity:
          type: string
          example: complex
        actors:
          type: array
          items:
            type: string
          example: ["Patient", "Doctor", "Admin"]
        features:
          type: array
          items:
            type: string
          example: ["Book Appointment", "Process Payments", "Cancel Schedule"]
        entities:
          type: array
          items:
            type: string
          example: ["PatientProfile", "DoctorSchedule", "AppointmentTransaction"]
        userId:
          type: string
          format: uuid
          example: fdb1e0a8-b648-4cb2-87c2-9e8cbb034f71
        createdAt:
          type: string
          format: date-time
          example: 2026-07-01T15:28:56.000Z
    Specification:
      type: object
      properties:
        id:
          type: integer
          example: 8
        projectId:
          type: integer
          example: 12
        masterJson:
          $ref: '#/components/schemas/MasterJson'
        approvalStatus:
          type: string
          enum: [pending, approved]
          example: pending
        srsMarkdown:
          type: string
          nullable: true
          example: null
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    MasterJson:
      type: object
      required:
        - project
        - actors
        - features
        - entities
        - attributes
        - relationships
      properties:
        project:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
            description:
              type: string
            domain:
              type: string
            complexity:
              type: string
        actors:
          type: array
          items:
            type: string
        features:
          type: array
          items:
            type: string
        entities:
          type: array
          items:
            type: string
        attributes:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          example:
            PatientProfile: ["profileId", "dateOfBirth", "bloodType"]
            DoctorSchedule: ["scheduleId", "workingHours", "isAvailable"]
        relationships:
          type: array
          items:
            $ref: '#/components/schemas/Relationship'
        diagrams:
          type: object
          nullable: true
          properties:
            usecase:
              $ref: '#/components/schemas/DiagramData'
            er:
              $ref: '#/components/schemas/DiagramData'
            class:
              $ref: '#/components/schemas/DiagramData'
    Relationship:
      type: object
      required:
        - source
        - target
        - type
        - label
      properties:
        source:
          type: string
          example: PatientProfile
        target:
          type: string
          example: AppointmentTransaction
        type:
          type: string
          enum: [one-to-one, one-to-many, many-to-many]
          example: one-to-many
        label:
          type: string
          example: books
    DiagramData:
      type: object
      properties:
        plantuml:
          type: string
          example: |
            @startuml
            actor Patient as patient
            ...
            @enduml
        url:
          type: string
          example: http://www.plantuml.com/plantuml/png/SyfFMieGuz98...
    CreateProjectRequest:
      type: object
      required:
        - name
        - description
        - domain
        - complexity
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 100
        description:
          type: string
          minLength: 10
          maxLength: 2000
        domain:
          type: string
          minLength: 2
          maxLength: 100
        complexity:
          type: string
          enum: [simple, standard, complex]
        actors:
          type: array
          items:
            type: string
        features:
          type: array
          items:
            type: string
        entities:
          type: array
          items:
            type: string
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Validation failed
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
          example: [{"field": "name", "message": "Must be at least 3 characters"}]